+++ PgRoutiner/Builder/CodeBuilder/Crud/CrudCreateOnConflictDoUpdateReturningCode.cs
@@ -80,7 +80,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I5}(\"{p.Name}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}();");
+            Class.AppendLine($"{I4}.{settings.ReturnMethod}();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, true);
         }
@@ -102,7 +102,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I5}(\"{p.Name}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I4}.{settings.SingleLinqMethod}Async();");
+            Class.AppendLine($"{I4}.{settings.ReturnMethod}Async();");
             Class.AppendLine($"{I2}}}");
             AddMethod(name, actualReturns, false);
         }
@@ -122,7 +122,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I4}(\"{p.Name}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}();");
+            Class.AppendLine($"{I3}.{settings.ReturnMethod}();");
             AddMethod(name, actualReturns, true);
         }
 
@@ -142,7 +142,7 @@ namespace PgRoutiner
             Class.AppendLine(", ");
             Class.Append(string.Join($",{NL}", this.ColumnParams.Select(p => $"{I4}(\"{p.PgName}\", model.{p.ClassName}, {p.DbType})")));
             Class.AppendLine($")");
-            Class.AppendLine($"{I3}.{settings.SingleLinqMethod}Async();");
+            Class.AppendLine($"{I3}.{settings.ReturnMethod}Async();");
             AddMethod(name, actualReturns, false);
         }