+++ PgRoutiner/Settings/ParseSettings.cs
@@ -157,9 +157,9 @@ namespace PgRoutiner
                 Value.Markdown = false;
             }
 
-            if (!new string[] { "Single", "SingleOrDefault", "First", "FirstOrDefault" }.Contains(Value.SingleLinqMethod))
+            if (!new string[] { "Single", "SingleOrDefault", "First", "FirstOrDefault" }.Contains(Value.ReturnMethod))
             {
-                Program.DumpError($"SingleLinqMethod setting must be one of the allowed values: Single, SingleOrDefault, First or FirstOrDefault");
+                Program.DumpError($"ReturnMethod setting must be one of the allowed values: Single, SingleOrDefault, First or FirstOrDefault");
                 return null;
             }